add validation to student form

+++ database/migrations/2023_05_08_120657_create_student_types_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('student_types', function (Blueprint $table) {
             $table->id();
             $table->string('type');
-            $table->float('dpp')->nullable();
-            $table->float('krs')->nullable();
-            $table->float('uts')->nullable();
-            $table->float('uas')->nullable();
-            $table->float('wisuda')->nullable();
+            $table->decimal('dpp', 10, 2)->nullable();
+            $table->decimal('krs', 10, 2)->nullable();
+            $table->decimal('uts', 10, 2)->nullable();
+            $table->decimal('uas', 10, 2)->nullable();
+            $table->decimal('wisuda', 10, 2)->nullable();
             $table->timestamps();
         });
     }
